refactor: move pod-spec helpers from support/util to support/podspec
Move containers.go, volumes.go, volumemounts.go, cloudprovidercreds.go,
deployment.go, and manifests.go (plus test files) into a new
support/podspec package.

Renames to avoid stuttering in the new package:
- PodVolumeMounts → VolumeMounts
- ContainerVolumeMounts → ContainerMounts
