+++ support/podspec/containers.go
@@ -1,4 +1,4 @@
-package util
+package podspec
 
 import (
 	"fmt"
