+++ support/podspec/deployment.go
@@ -1,4 +1,4 @@
-package util
+package podspec
 
 import (
 	"context"
