+++ support/podspec/deployment_test.go
@@ -1,4 +1,4 @@
-package util
+package podspec
 
 import (
 	"testing"
