+++ support/podspec/manifests.go
@@ -1,4 +1,4 @@
-package util
+package podspec
 
 import (
 	corev1 "k8s.io/api/core/v1"
