+++ support/podspec/volumes.go
@@ -1,4 +1,4 @@
-package util
+package podspec
 
 import (
 	appsv1 "k8s.io/api/apps/v1"
